tempo

+++ conversion/apply_modifier_nbr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   apply_modifier_nbr.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fkante <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: fkante <fkante@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/03 17:33:08 by fkante            #+#    #+#             */
-/*   Updated: 2019/09/04 12:16:22 by fkante           ###   ########.fr       */
+/*   Updated: 2019/09/04 14:31:29 by amartino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int64_t		apply_modifier_di(int64_t nbr_conv, uint64_t flag)
+int64_t		apply_modifier_di(int64_t nbr_conv, uint32_t flag)
 {
 	if (flag & FLAG_HH)
 		nbr_conv = (int8_t)nbr_conv;
@@ -25,7 +25,7 @@ int64_t		apply_modifier_di(int64_t nbr_conv, uint64_t flag)
 	return (nbr_conv);
 }
 
-uint64_t	apply_modifier_oxX(int64_t nbr_conv, uint64_t flag)
+uint64_t	apply_modifier_oxX(int64_t nbr_conv, uint32_t flag)
 {
 	if (flag & FLAG_HH)
 		nbr_conv = (uint8_t)nbr_conv;
@@ -38,7 +38,7 @@ uint64_t	apply_modifier_oxX(int64_t nbr_conv, uint64_t flag)
 	return (nbr_conv);
 }
 
-uint64_t	apply_modifier_u(uint64_t nbr_conv, uint64_t flag)
+uint64_t	apply_modifier_u(uint64_t nbr_conv, uint32_t flag)
 {
 	if (flag & FLAG_HH)
 		nbr_conv = (uint8_t)nbr_conv;
